+++ packages/hooks/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @codeward/hooks
+# @codeforward/hooks
 
 ## 1.0.0